fix: 중복 클래스 제거, 로직 통합
