--- ZipCock_44/src/main/java/org/danpoong/zipcock_44/domain/user/User.java
@@ -1,38 +1,0 @@
-package org.danpoong.zipcock_44.domain.user;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.danpoong.zipcock_44.global.common.BaseTimeEntity;
-import org.danpoong.zipcock_44.domain.chat.entity.ChatRoom;
-import org.danpoong.zipcock_44.domain.post.entity.Post;
-import org.danpoong.zipcock_44.global.common.BaseTimeEntity;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "USER")
-public class User extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
-    private long id;
-
-    @Column(name = "user_name")
-    private String name;
-
-    @OneToMany(mappedBy = "user")
-    private List<Post> posts = new ArrayList<>();
-
-    @OneToMany(mappedBy = "buyer", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<ChatRoom> buyerChatRooms = new ArrayList<>();
-
-    @OneToMany(mappedBy = "seller", cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<ChatRoom> sellerChatRooms = new ArrayList<>();
-
-}
