+++ ZipCock_44/src/main/java/org/danpoong/zipcock_44/domain/user/service/KakaoService.java
@@ -1,28 +1,22 @@
 package org.danpoong.zipcock_44.domain.user.service;
 
 import io.netty.handler.codec.http.HttpHeaderValues;
-import jakarta.servlet.http.HttpSession;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.danpoong.zipcock_44.domain.user.UserRepository;
 import org.danpoong.zipcock_44.domain.user.dto.response.KakaoLogoutResponseDTO;
 import org.danpoong.zipcock_44.domain.user.dto.response.KakaoTokenResponseDTO;
 import org.danpoong.zipcock_44.domain.user.dto.response.KakaoUserInfoResponseDTO;
 import org.danpoong.zipcock_44.domain.user.entity.User;
+import org.danpoong.zipcock_44.domain.user.repository.UserRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
 import org.springframework.http.*;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
